Refactor: base resubmitting publisher & dlq listener
